requirements.txt
streamlit
torch
torchvision
Pillow
tensorflow